# XXH32a/XXH64a: XXH32, vectorized