--- a/Makefile
+++ b/Makefile
@@ -198,7 +198,7 @@ staticAnalyze: clean
 .PHONY: cppcheck
 cppcheck:
 	@echo ---- static analyzer - cppcheck ----
-	cppcheck . --force --enable=warning,portability,performance,style --error-exitcode=1 --suppress=knownConditionTrueFalse > /dev/null
+	cppcheck . --force --enable=warning,portability,performance,style --inline-suppr --template='{callstack}:{file}:{line},{severity},{id},{message}' --error-exitcode=1 --suppress=knownConditionTrueFalse #> /dev/null
 
 .PHONY: namespaceTest
 namespaceTest: