--- a/xxhash-vec.h
+++ b/xxhash-vec.h
@@ -36,9 +36,11 @@
 #define XXHASH_VEC_H
 
 #if defined(__ARM_NEON__) || defined(__ARM_NEON)
-#  include <arm_neon.h>
-#  define XXH_NEON
-#  define XXH_VECTORIZE 1
+#include <arm_neon.h>
+#define XXH_NEON
+#undef XXH_VECTORIZE
+#define XXH_VECTORIZE 1
+
 typedef uint32x4_t U32x4;
 typedef uint32x4x2_t U32x4x2;
 
@@ -68,7 +70,8 @@ FORCE_INLINE U32x4 XXH_rotlvec_vec32(U32x4 x, const U32x4 r)
 
 #elif (XXH_GCC_VERSION >= 407 || defined(__clang__)) \
 	&& (defined(__SSE4_1__) || defined(__AVX__) || defined(_M_X64) || defined(_M_IX86_FP))
-
+#undef XXH_VECTORIZE
+#define XXH_VECTORIZE 1
 /* not NEON */
 /* __m128i (SSE) or uint32x4_t (NEON). */
 typedef U32 U32x4 __attribute__((__vector_size__(16)));
@@ -113,6 +116,7 @@ FORCE_INLINE void XXH_vec_store_unaligned(void* p, const U32x4 v)
 
 /* This catches MSVC++ if supplied /TP, and hopefully ICC. */
 #elif defined(__cplusplus) && (defined(__SSE4_1__) || defined(__AVX__) || defined(_M_X64) || defined(_M_IX86_FP))
+#undef XXH_VECTORIZE
 #define XXH_VECTORIZE 1
 #include <smmintrin.h>