… README.md
… xxhash.h
… xxhsum.c