This uses the tips to vectorize xxHash32 from https://moinakg.wordpress.com/2013/01/19/vectorizing-xxhash-for-fun-and-profit/ but without the x86 intrinsics. Instead, it uses GCC vector extensions to implement the same behavior naturally.

It supports (presumably) all versions of Clang, and GCC 4.7+, as long as they target SSE4.1 (`-march=core2`, `-march=penryn`, `-msse4.1`), or ARMv7-a with NEON.

It changes the hash which sucks a lot, but the performance seems very promising once it gets tweaked a bit.

Aligned 32-bit hashes on my MacBook (2.13 GHz Core 2 Duo/P7450/Penryn) have a 1.1-1.16x speedup, but a yucky 0.8-0.9x slowdown for unaligned hashes.

Before (`gcc-8 -mno-sse4`):
```
XXH32               :     102400 ->    40398 it/s ( 3945.1 MB/s)
XXH32 unaligned     :     102400 ->    31806 it/s ( 3106.1 MB/s)
XXH64               :     102400 ->    41097 it/s ( 4013.3 MB/s)
XXH64 unaligned     :     102400 ->    29493 it/s ( 2880.2 MB/s)
```

After (`gcc-8 -ftree-vectorize -ftree-slp-vectorize -msse4.1`):
```
XXH32               :     102400 ->    47076 it/s ( 4597.3 MB/s)
XXH32 unaligned     :     102400 ->    29567 it/s ( 2887.4 MB/s)
XXH64               :     102400 ->    41151 it/s ( 4018.7 MB/s)
XXH64 unaligned     :     102400 ->    29494 it/s ( 2880.3 MB/s)
```
This actually makes 32-bit hashes __faster__ than 64-bit on my machine.

~~I still need to fix update_endian and make unaligned and 64-bit faster, and I want to know how this performs on AVX (Sandy Bridge/2nd Gen Core i-series, use `-mavx`).~~ AVX performs fine here and endianness is taken care of.

In addition, I just added XXH64a.  

The difference between XXH32a and XXH64a are minimal. They use the exact same inner loops and state type (they are treated separately, but typedef'd to each other) but the beginning and end are handled differently.

XXH32a will copy the seed and merge the lanes in the digest, while XXH64a will split the seed between the 8 lanes and join them together.

The result of XXH64a is a 64-bit hash that has good performance on both 32-bit and 64-bit systems.

XXH64 will (usually) be much faster on 64-bit (excluding pre-2010 Intel chips), but you will not see the drastic slowdown from software arithmetic if your code is run on 32-bit.