@easyaspi314 2018-11-22
Note: `xxhsum -b` will fail because of the mismatched checksum. Commenting out the code in `BMK_sanityCheck` will force it to run.