@easyaspi314 2018-11-22
LG G3 results are a little inconsistent to say the least.

3 consecutive runs (leaving out the obviously terrible XXH64):
```
XXH32               :     102400 ->    38025 it/s ( 3713.4 MB/s)
XXH32 unaligned     :     102400 ->    44340 it/s ( 4330.1 MB/s)
XXH64               :     102400 ->     9093 it/s (  888.0 MB/s)
XXH64 unaligned     :     102400 ->     9093 it/s (  888.0 MB/s)

XXH32               :     102400 ->    53669 it/s ( 5241.1 MB/s)
XXH32 unaligned     :     102400 ->    54507 it/s` ( 5323.0 MB/s)

XXH32               :     102400 ->    37528 it/s ( 3664.9 MB/s)
XXH32 unaligned     :     102400 ->    38100 it/s ( 3720.7 MB/s)
```
However, it is still a lot faster than vanilla:
```
XXH32               :     102400 ->    23829 it/s ( 2327.1 MB/s)
XXH32 unaligned     :     102400 ->    23872 it/s ( 2331.2 MB/s)
```