@Cyan4973 2018-11-27
That's great work @easyaspi314 !

It's a nice variant, and the learnings are definitely interesting.
The ARM target seems to benefit greatly.

Obviously, a major issue is that if it generates a different hash value, it cannot be branded `XXH32`, otherwise it would introduce confusion.

Some potential follow up : 
- Would it be possible to use the same methodology while preserving original result ?
    + It seems that, on ARM, vectorization may end up being beneficial compared to scalar
    + On x86, it seems a worse bet. Maybe try to use two 64-bit vectors, instead of a single 128-bit one, in order to interleave some operations ?

- Regarding the non-compatible output variant : 
    + how does it behave presuming vectorization is disabled ?

- Question : how come the speed of `XXH32` and `XXH64` seem similar on your Penryn platform ?
I must admit I'm very surprised, `XXH64` has always proved much faster (aka ~2x faster) than `XXH32` on AMD64 instruction set, even on my old penryn system (now decommissioned unfortunately). Seeing them equivalent is quite a surprise.