@easyaspi314 2018-11-27
How about we call it `XXH32x2`?