@Cyan4973 2018-11-27
> GCC is a lot more conservative when it comes to vectorizing loops, and it won't do it with `-O2`. You need `-O3`.

I tested `gcc` with `-O3 -mavx2`.
The observed gain is negligible, within noise level (9.5 GB/s => 9.6 GB/s).
When disabling vectorization `-O3 -mno-sse4`, the speed loss is visible : down to 7 GB/s.
So `gcc` code was likely vectorized with `-mavx2`.
But for some reason, `gcc`'s generated code is slower than `clang`'s.