@easyaspi314 2018-11-28
XXH32 vectorized is now XXH32a, and tests are updated.

I also improved performance on classic Thumb-1 by forcing GCC to emit a multiply instruction instead of [this](https://godbolt.org/z/2vc82a), and making the normal hash functions check alignment and endianness first and use a simpler version with less branching.