@Cyan4973 2018-11-28
Any performance figure available on ARM target ?

I tested it on a modern Core9, 
results are similar as previously : 
`clang` gets better speed on `XXH32a` when vectorization and `sse4` are enabled,
but it also craters `XXH32` because it tries to auto-vectorize it, and that's a wrong bet.
`gcc` performance, on the other hand, doesn't change at all,
whether it is for `XXH32` or `XXH32a`, with or without vectorization / `sse4` / `avx2` enabled.

So I still believe that `ARM` is a better target for this approach.
Although I have no idea what's the state of vectorization support for ARM, given the large range of cpu variants.