@easyaspi314 2018-11-28
Yes, I agree. Vectorization, especially with the vsli/vshr hint, is incredibly easy on ARM, thanks to its simple syntax, RISC nature, and similarity to the rest of the instruction set.

So far, since we have the benchmark of XXH32 next to XXH32a, NEON XXH32a is ~2.2x-2.5x faster than XXH32.  I still need to figure out the correct setting to make a stable benchmark.

Unfortunately, I have been wrestling with the ARMv7r set (basically ARMv7a minus NEON). XXH32a ends up being much slower than XXH32, even though on ARMv4t ARM mode, they are about the same.

As for auto-vectorization, it is a very touchy subject for compilers. As we are seeing, XXH32 doesn't really like autovec, and slows down significantly, and GCC is much more picky than Clang, presumably because Clang was designed with vectorization in mind, while GCC was based in the 80s-90s.