@easyaspi314 2018-11-28
Here is vanilla on my phone
```
XXH32               :     102400 ->    23712 it/s ( 2315.6 MB/s)
XXH32 unaligned     :     102400 ->    23824 it/s ( 2326.6 MB/s)
XXH64               :     102400 ->     9043 it/s (  883.1 MB/s)
XXH64 unaligned     :     102400 ->     9041 it/s (  882.9 MB/s)
```