@Cyan4973 2018-11-28
You are right, the NEON version of XXH32 is really good, and worth being integrated.

Do you believe you could extract the XXH32 part of your patch, and make it a dedicated PR ?
I believe it can be merged into mainline, provided NEON detection is automated and doesn't conflict with other targets, which seems to be the case. I would like to review it.