@easyaspi314 2018-11-28
Ooh I found a little treat that can fix our x86_64 problem!

```c
#ifndef __has_attribute
#define __has_attribute(x) 0
#endif

#if !defined(XXH_FORCE_VECTORIZE) && defined(__SSE4_1__)
# if XXH_GCC_VERSION >= 404
#  define XXH_SSE_OFF _Pragma("GCC push_options") _Pragma("GCC target(\"no-sse4\")")
#  define XXH_SSE_ON _Pragma("GCC pop_options")
# endif 
# if defined(__clang__) && __has_attribute(__target__)
#  define NO_SSE __attribute__((__target__("no-sse4")))
# endif 
#endif 
#ifndef XXH_SSE_OFF
# define XXH_SSE_OFF
# define XXH_SSE_ON
#endif 
#ifndef NO_SSE
# define NO_SSE
#endif 

XXH_SSE_OFF
/* XXH32 stuff here. */

NO_SSE XXH_PUBLIC_API unsigned int XXH32 (const void* input, size_t len, unsigned int seed)
{...}
XXH_SSE_ON

U32 XXH32a(...);
```

Now, on supported GCC and Clang versions, XXH32 will no longer compile with SSE4.1 instructions.